pyproject.toml
[tool.poetry]
name = "rubix-lorawan"
version = "0.1.0"
description = ""
authors = ["NubeDev <aidan@enviroservices.com.au>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
toml = "^0.10.2"
schedule = "^1.1.0"
gunicorn = "^20.1.0"
requests = "^2.25.1"
rubix-registry = {git = "https://github.com/NubeIO/rubix-registry", rev = "v1.0.0"}
rubix-http = {git = "https://github.com/NubeIO/rubix-http", rev = "v1.1.1"}
rubix-mqtt = {git = "https://github.com/NubeIO/rubix-mqtt", rev = "v1.2.0"}

SQLAlchemy = "^1.4.14"
paho-mqtt = "^1.5.1"
gevent = "^21.1.2"
py-chirpstack = {git = "https://github.com/NubeIO/py-chirpstack"}

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
